Remove hardcoded server ids in checking for admin and moderator rights
This will now check for the supported Discord ADMINISTRATOR permission
instead of checking for the role name Admin. The 'moderator' role name
still needs to be hard-coded though.
