+++ src/utils/index.ts
@@ -1,5 +1,5 @@
 export * from './isBlank';
-export * from './isMessageSentFromAdmin';
+export * from './isSentFromAdmin';
 export * from './messageFetcher';
 export * from './messageProcessor';
 export * from './random';
